Binpkg: fix _chpathtool_exit rmtree ENOENT after rename

+++ lib/_emerge/Binpkg.py
@@ -378,14 +378,15 @@ class Binpkg(CompositeTask):
 		image_tmp_dir = os.path.join(
 			self.settings["PORTAGE_BUILDDIR"], "image_tmp")
 		build_d = os.path.join(self.settings["D"],
-			self._build_prefix.lstrip(os.sep))
+			self._build_prefix.lstrip(os.sep)).rstrip(os.sep)
 		if not os.path.isdir(build_d):
 			# Assume this is a virtual package or something.
 			shutil.rmtree(self._image_dir)
 			ensure_dirs(self.settings["ED"])
 		else:
 			os.rename(build_d, image_tmp_dir)
-			shutil.rmtree(self._image_dir)
+			if build_d != self._image_dir:
+				shutil.rmtree(self._image_dir)
 			ensure_dirs(os.path.dirname(self.settings["ED"].rstrip(os.sep)))
 			os.rename(image_tmp_dir, self.settings["ED"])
